% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_category.R
\name{category_delete}
\alias{category_delete}
\title{Delete a phenotype category}
\usage{
category_delete(name, conn)
}
\arguments{
\item{name}{phenotype category name in the project}

\item{conn}{gor connection structure, create it using \code{\link{phenotype_connect}} or \code{\link{gor_connect}}}
}
\description{
Delete a phenotype category
}
\examples{
\dontrun{
api_key <- Sys.getenv("GOR_API_KEY")
project <- Sys.getenv("GOR_PROJECT")
conn <- phenotype_connect(api_key, project)
name <- "testCat"
category_delete(name, conn)
}
}
