% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_category.R
\name{create_category}
\alias{create_category}
\title{Add a new category to this project.}
\usage{
create_category(name, conn)
}
\arguments{
\item{name}{Unique (lowercase) category name in the project}

\item{conn}{gor connection structure, create it using \code{\link{phenotype_connect}} or \code{\link{gor_connect}}}
}
\value{
a list of category attrbues
}
\description{
Add a new category to this project.
}
\examples{
\dontrun{
api_key <- Sys.getenv("GOR_API_KEY")
project <- Sys.getenv("GOR_PROJECT")
conn <- phenotype_connect(api_key, project)
name <- "testCat1"
category <- create_category(name, conn)
}
}
