% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype-catalog_services.R
\name{create_phenotype}
\alias{create_phenotype}
\title{Create a new phenotype in the current project.}
\usage{
create_phenotype(
  name,
  result_type,
  conn,
  description = NULL,
  url = NULL,
  category = NULL,
  query = NULL,
  tags = NULL,
  data = NULL
)
}
\arguments{
\item{name}{Unique (lowercase) phenotype name in the project}

\item{result_type}{Type of phenotype (supported types: "SET", "QT" and "CATEGORY")}

\item{conn}{gor connection structure, create it using \code{\link{platform_connect}}}

\item{description}{Optional Phenotype description}

\item{url}{Reference URL for the phenotype (to dataset or other reference)}

\item{category}{Enter the category for the phenotype (must be defined in the project - see get_categories) (optional)}

\item{query}{NOR query that defines this phenotype (optional)}

\item{tags}{comma separated string of tags eg. "height,weight" or character vector "c("height", "weight") (optional)}

\item{data}{data.frame or list of phenotype PN values to be applied to \code{phenotype_upload_data} (optional)}
}
\value{
a list with the phenotype object
}
\description{
Create a new phenotype in the current project.
}
\examples{
\dontrun{
api_key <- Sys.getenv("GOR_API_KEY")
project <- Sys.getenv("GOR_PROJECT")
conn <- platform_connect(api_key, project)
name <- "height"
result_type <- "QT"
description <- "Height of individuals"
phenotype <- create_phenotype(name, result_type, conn, description)
}
}
