% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype-catalog_services.R
\name{create_playlist}
\alias{create_playlist}
\title{Create a new playlist in the current project.}
\usage{
create_playlist(name, conn, description = NULL, phenotypes = NULL)
}
\arguments{
\item{name}{Unique (lowercase) playlist name in the project}

\item{conn}{gor connection structure, create it using \code{\link{platform_connect}}}

\item{description}{Free text description of the playlist (optional)}

\item{phenotypes}{comma seperated string of phenotypes to  add (optional) eg. "pheno1,pheno2"  or character vector c("pheno1", "pheno2")}
}
\value{
A playlist object
}
\description{
Create a new playlist in the current project.
}
\examples{
\dontrun{
api_key <- Sys.getenv("GOR_API_KEY")
project <- Sys.getenv("GOR_PROJECT")
conn <- platform_connect(api_key, project)
name <- "testpl"
playlist <- create_playlist(name = name, conn)
}
}
