% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_access_token.R
\name{get_access_token}
\alias{get_access_token}
\title{Get an access token from a refresh token}
\usage{
get_access_token(api_key, url)
}
\arguments{
\item{api_key}{the api key}

\item{url}{the full url to the authentication service}
}
\value{
returns http headers, constructed with \code{\link[httr]{add_headers}}
}
\description{
Get an access token from a refresh token
}
