% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-catalog_services.R
\name{get_analysis_catalog_run}
\alias{get_analysis_catalog_run}
\title{Get an analysis catalog run in the current project by name.}
\usage{
get_analysis_catalog_run(
  analysis_catalog_name,
  analysis_catalog_run_name,
  conn
)
}
\arguments{
\item{analysis_catalog_name}{The name of the Analysis Catalog}

\item{analysis_catalog_run_name}{The name of the Analysis Catalog run}

\item{conn}{gor connection structure, create it using \code{\link{platform_connect}}}
}
\value{
analysis_catalog_run structure
}
\description{
Get an analysis catalog run in the current project by name.
}
\examples{
\dontrun{
api_key <- Sys.getenv("GOR_API_KEY")
project <- Sys.getenv("GOR_PROJECT")
conn <- platform_connect(api_key, project)
catalog <- get_analysis_catalog_run(analysis_catalog_name="test_ac_1",
                                      analysis_catalog_run_name = "test_ac_run_1",
                                      conn = conn)
}
}
