% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-catalog_services.R
\name{get_analysis_catalog_runs}
\alias{get_analysis_catalog_runs}
\title{A list of all the analysis catalog runs in the current project for a given phenotype.}
\usage{
get_analysis_catalog_runs(phenotype_name, conn, limit = 100)
}
\arguments{
\item{phenotype_name}{Only list analysis catalog runs for a specific phenotype name}

\item{conn}{gor connection structure, create it using \code{\link{platform_connect}}}

\item{limit}{Maximum number of results (default: 100)}
}
\value{
List of analysis catalog runs
}
\description{
A list of all the analysis catalog runs in the current project for a given phenotype.
}
\examples{
\dontrun{
api_key <- Sys.getenv("GOR_API_KEY")
project <- Sys.getenv("GOR_PROJECT")
conn <- platform_connect(api_key, project)
catalog <- get_analysis_catalog_runs(phenotype_name = "test_pheno", conn = conn)
}
}
