% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-catalog_services.R
\name{get_analysis_catalogs}
\alias{get_analysis_catalogs}
\title{A list of all the analysis catalogs in the current project. Optionally scope results to a given phenotype name.}
\usage{
get_analysis_catalogs(conn, phenotype_name = NULL)
}
\arguments{
\item{conn}{gor connection structure, create it using \code{\link{platform_connect}}}

\item{phenotype_name}{optional: Only fetch analysis catalogs that relate to a given phenotype name}
}
\value{
List of analysis catalogs
}
\description{
A list of all the analysis catalogs in the current project. Optionally scope results to a given phenotype name.
}
\examples{
\dontrun{
api_key <- Sys.getenv("GOR_API_KEY")
project <- Sys.getenv("GOR_PROJECT")
conn <- platform_connect(api_key, project)
catalogs <- get_analysis_catalogs(conn, limit=5)
}
}
