% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_category.R
\name{get_categories}
\alias{get_categories}
\title{A list of all categories available in the system.}
\usage{
get_categories(conn)
}
\arguments{
\item{conn}{gor connection structure, create it using \code{\link{gor_connect}}}
}
\value{
List of categories
}
\description{
A list of all categories available in the system.
}
\examples{
\dontrun{
api_key <- Sys.getenv("GOR_API_KEY")
project <- Sys.getenv("GOR_PROJECT")
conn <- phenotype_connect(api_key, project)
phenotypes <- get_categories(conn)
}
}
