% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_services.R
\name{get_covariate}
\alias{get_covariate}
\title{Get a single covariate by its id.}
\usage{
get_covariate(id, conn)
}
\arguments{
\item{id}{Unique covariate id in the project}

\item{conn}{platform connection structure, create it using \code{\link{platform_connect}}}
}
\value{
a covariate
}
\description{
Get a single covariate by its id.
}
