% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_matrix_svc.R
\name{get_data}
\alias{get_data}
\title{Retrieve a phenotype matrix from the server.}
\usage{
get_data(phenotype_matrix, conn)
}
\arguments{
\item{phenotype_matrix}{phenotype_matrix structure.}

\item{conn}{gor connection structure, create it using
\code{\link{phenotype_connect}} or \code{\link{gor_connect}}}
}
\value{
phenotype matrix tibble from server
}
\description{
Can be called after phenotypes have been added to the request. Create it
using \code{\link{get_phenotype_matrix}} and add to it
using \code{\link{phemat_add_phenotype}} or \code{\link{phemat_add_phenotypes}}
}
\examples{
\dontrun{
api_key <- Sys.getenv("GOR_API_KEY")
project <- Sys.getenv("GOR_PROJECT")
conn <- phenotype_connect(api_key, project)
phenotype_mat <- get_phenotype_mat()
phenotype_mat <-  phemat_add_phenotypes(...)
phenotype_data <- get_data(phenotype_mat, conn)
}
}
