% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_services.R
\name{get_job}
\alias{get_job}
\title{Get a workflow job}
\usage{
get_job(job_id, conn)
}
\arguments{
\item{job_id}{job-id as integer or l}

\item{conn}{connection object, see \code{\link{platform_connect}}}
}
\value{
WorkflowJob object
}
\description{
This is a description
}
