% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_services.R
\name{get_jobs}
\alias{get_jobs}
\title{Get a list of jobs satisfying the supplied criteria}
\usage{
get_jobs(
  conn,
  user_name = NULL,
  status = NULL,
  project = NULL,
  pipeline = NULL,
  state = NULL,
  context = NULL,
  limit = 20
)
}
\arguments{
\item{conn}{connection object, see \code{\link{platform_connect}}}

\item{user_name}{The user who created the job}

\item{status}{Current status of jobs}

\item{project}{Filter by project}

\item{pipeline}{Filter by pipeline name}

\item{state}{Filter by state, each state encapsulates several statuses (running, finished)}

\item{context}{Filter by context string}

\item{limit}{Maximum number of results. Default 20}
}
\value{
data.frame of jobs
}
\description{
Get a list of jobs satisfying the supplied criteria
}
