% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_svc.R
\name{get_phenotype}
\alias{get_phenotype}
\title{Get a specific phenotype in the current project.}
\usage{
get_phenotype(name, conn)
}
\arguments{
\item{name}{Unique (lowercase) phenotype name in the project}

\item{conn}{gor connection structure, create it using \code{\link{phenotype_connect}} or  \code{\link{gor_connect}}}
}
\value{
a list with the phenotype object
}
\description{
Get a specific phenotype in the current project.
}
\examples{
\dontrun{
api_key <- Sys.getenv("GOR_API_KEY")
project <- Sys.getenv("GOR_PROJECT")
conn <- phenotype_connect(api_key, project)
name <- "height"
phenotype <- get_phenotype(name, conn)
}
}
