% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype-catalog_services.R
\name{get_phenotype_matrix}
\alias{get_phenotype_matrix}
\title{Get a phenotype matrix object.}
\usage{
get_phenotype_matrix(base = NULL, ...)
}
\arguments{
\item{base}{Optional name of base set}

\item{...}{named arguments passed to \code{get_phenotypes} for populating matrix.}
}
\value{
a phenotype matrix object
}
\description{
Get a phenotype matrix object.
}
\examples{
\dontrun{
phenotype_mat <- get_phenotype_matrix()
}
}
