% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_get_data.R
\name{get_phenotypes_data}
\alias{get_phenotypes_data}
\title{Retrieve phenotype data from the server.}
\usage{
get_phenotypes_data(pheno_names, conn, missing_value = NULL, base = NULL)
}
\arguments{
\item{pheno_names}{Fetch phenotypes that are in a given list of phenotype names, character and character vector also allowed}

\item{conn}{platform connection structure, create it using \code{\link{platform_connect}}}

\item{missing_value}{The string to substitute for a missing value in the data}

\item{base}{Optional name of base set}
}
\value{
tibble from server
}
\description{
Retrieve phenotype data from the server.
}
