% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_services.R
\name{get_phenotypes_dataframe}
\alias{get_phenotypes_dataframe}
\title{A dataframe of all/subset of phenotypes in the current project.}
\usage{
get_phenotypes_dataframe(
  conn,
  pheno_names = list(),
  all_tags = list(),
  any_tags = list(),
  pn_count = NULL,
  categories = list(),
  limit = 100,
  states = list(),
  search = NULL,
  playlist = NULL,
  updated_at = NULL,
  result_types = list(),
  filtered = TRUE
)
}
\arguments{
\item{conn}{gor connection structure, create it using \code{\link{platform_connect}}}

\item{pheno_names}{Only fetch phenotypes that are in a given list of phenotype names}

\item{all_tags}{Only fetch phenotypes that have all tags in the given list of tags}

\item{any_tags}{Fetch phenotypes that have any of the tags in the given list of tags}

\item{pn_count}{Only list phenotypes that match the given pn counts. (include greater or less symbol in string eg ">10")}

\item{categories}{Only fetch phenotypes in the given list of categories}

\item{limit}{Maximum number of results (default: 100)}

\item{states}{Only fetch phenotypes in the given list of states}

\item{search}{String of keywords to search for in phenotypes, such as name, categories and tags}

\item{playlist}{Fetch a specific playlist of phenotypes by the playlist id}

\item{updated_at}{Only fetch phenotypes that match the given dates. Example: >=2017-04-01 ┃ <=2012-07-04 ┃ 2016-04-30..2016-07-04}

\item{result_types}{Only fetch phenotypes in the given list of result types}

\item{filtered}{boolean - Return subset of phenotype info columns. filtered=FALSE returns everything. Default: TRUE}
}
\value{
Dataframe of phenotype/s info
}
\description{
A dataframe of all/subset of phenotypes in the current project.
}
\examples{
\dontrun{
api_key <- Sys.getenv("GOR_API_KEY")
project <- Sys.getenv("GOR_PROJECT")
conn <- platform_connect(api_key, project)
phenotypes <- get_phenotypes_dataframe(conn, limit=5)
}
}
