% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_services.R
\name{get_pipelines}
\alias{get_pipelines}
\title{Returns the pipelines available on the current server}
\usage{
get_pipelines(conn, include.description = FALSE)
}
\arguments{
\item{conn}{connection object, see \code{\link{platform_connect}}}

\item{include.description}{Logical, should pipeline description be included in data.frame results. Default: False}
}
\value{
data.frame oor list of pipeline info
}
\description{
Refer to the API documentation for the Workflow service to see formatting of data.
}
