% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_playlist.R
\name{get_playlist}
\alias{get_playlist}
\title{Get playlist in the current project based on the playlist id.}
\usage{
get_playlist(id, conn)
}
\arguments{
\item{id}{Playlist id}

\item{conn}{gor connection structure, create it using \code{\link{platform_connect}}}
}
\value{
A playlist object
}
\description{
Get playlist in the current project based on the playlist id.
}
\examples{
\dontrun{
api_key <- Sys.getenv("GOR_API_KEY")
project <- Sys.getenv("GOR_PROJECT")
conn <- platform_connect(api_key, project)
playlist <- get_playlist(id=1, conn)
}
}
