% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_playlist.R
\name{get_playlists}
\alias{get_playlists}
\title{A list of all the playlists in the current project.}
\usage{
get_playlists(conn, limit = 100)
}
\arguments{
\item{conn}{gor connection structure, create it using \code{\link{platform_connect}}}

\item{limit}{Maximum number of results (default: 100)}
}
\value{
List of playlists
}
\description{
A list of all the playlists in the current project.
}
\examples{
\dontrun{
api_key <- Sys.getenv("GOR_API_KEY")
project <- Sys.getenv("GOR_PROJECT")
conn <- platform_connect(api_key, project)
playlists <- get_playlists(conn)
}
}
