% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_tag.R
\name{get_tags}
\alias{get_tags}
\title{Get all tags}
\usage{
get_tags(conn)
}
\arguments{
\item{conn}{gor connection structure, create it using \code{\link{phenotype_connect}} or \code{\link{gor_connect}}}
}
\value{
A list of phenotype tags
}
\description{
Get all tags available in the catalog
}
\examples{
\dontrun{
api_key <- Sys.getenv("GOR_API_KEY")
project <- Sys.getenv("GOR_PROJECT")
conn <- phenotype_connect(api_key, project)
tags <- get_tags(conn)
}
}
