% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gor_connect.R
\name{gor_connect}
\alias{gor_connect}
\title{Set up a connection object, and call the health endpoint to make sure everything is up and running on the
server side}
\usage{
gor_connect(api_key = NULL, project = NULL,
  api_endpoint = "/api/query")
}
\arguments{
\item{api_key}{the api key from the /api-key-service endpoint of your CSA host through a browser}

\item{project}{project name}

\item{api_endpoint}{the api endpoint path on the server, defaults to /api/query}
}
\value{
returns a list with the connection data
}
\description{
Set up a connection object, and call the health endpoint to make sure everything is up and running on the
server side
}
\examples{
\dontrun{
api_key <- "...paste from your /api-key-service/token endpoint..."
conn <- gor_connect(api_key, "test_project")
}
}
