% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gor_create.R
\name{gor_create}
\alias{gor_create}
\title{gor_create}
\usage{
gor_create(..., conn = NULL)
}
\arguments{
\item{...}{ellipse for the relations to include}

\item{conn}{connection object from \code{\link{gor_connect}}}
}
\value{
partial-application of function \code{\link{gor_query}} with the \code{conn} and \code{relations} parameters set.
}
\description{
Creates a partial-application of \code{\link{gor_query}} in the context of \code{conn} and the listed relations in the \code{...} arguments
}
\details{
Relations can be either local dataframes (uploaded) or string queries(turned into create statements and prepended to the query when executed).
}
\examples{
\dontrun{
  query <- gor_create(air = airquality, conn = conn)
  query("nor [air] | where Treatment = 'nonchilled'")

}
}
