% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gor_define.R
\name{gor_define}
\alias{gor_define}
\title{GOR Define}
\usage{
gor_define(x)
}
\arguments{
\item{x}{can be a string, \code{gor_definitions} structure, list, or NULL}
}
\description{
Takes a parameter and parses it into a \code{gor_definitions} structure, which is basically a named list
of define statements, i.e. \verb{def name = code;}
}
\examples{
# below are just examples of input, not necessarily valid gor `def` code
gor_define("
   def #var1# = 10;
   def fun($1, $2) = gor $1
       | some
       | pipe
       | step $2;
")

gor_define(list(`#var1#` = 10, limit = 20, block = "| some | pipe | step"))
}
