% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gor_knitr_setup.R
\name{gor_knitr_setup}
\alias{gor_knitr_setup}
\title{GOR Knitr Setup}
\usage{
gor_knitr_setup()
}
\value{
none
}
\description{
Sets up 'gor' as a supported engine for knitr
}
\examples{
\dontrun{
gor_knitr_setup()
}
}
