% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gor_query.R
\name{gorr__cat}
\alias{gorr__cat}
\title{Custom wrapper for cli::cat..()}
\usage{
gorr__cat(line = T, rule = F, col = "white")
}
\arguments{
\item{line}{logical. Start with newline. Default True}

\item{rule}{logical. Include rule. Default False}

\item{col}{string, color of rule. Default 'white'}
}
\description{
Custom wrapper for cli::cat..()
}
