% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gor_query.R
\name{gorr__elapsed_time}
\alias{gorr__elapsed_time}
\title{Custom wrapper for formatted query progress messages}
\usage{
gorr__elapsed_time(elapsed, status = "", info = "")
}
\arguments{
\item{elapsed}{elapsed time}

\item{status}{optional query status message}

\item{info}{optional additional info message}
}
\description{
Custom wrapper for formatted query progress messages
}
