% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gor_query.R
\name{gorr__failure}
\alias{gorr__failure}
\title{Custom wrapper for stop() with formated error messages}
\usage{
gorr__failure(msg, detail = NULL, url = NULL)
}
\arguments{
\item{msg}{exception message}

\item{detail}{exception details (chr or chr vector)}

\item{url}{query url}
}
\description{
Custom wrapper for stop() with formated error messages
}
