% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gor_query.R
\name{gorr__get_query_results}
\alias{gorr__get_query_results}
\title{Get server query results}
\usage{
gorr__get_query_results(
  query_result_url,
  conn,
  spinner = invisible,
  query_limit = 10000,
  offset = 0,
  parse = T
)
}
\arguments{
\item{query_result_url}{query result url}

\item{conn}{connection object, see \code{\link{gor_connect}}}

\item{spinner}{spinner function (e.g. some ascii animation for long running queries)}

\item{query_limit}{limit to how many rows to fetch in each page}

\item{offset}{query offset (skip rows)}

\item{parse}{should output be parsed into a data.frame}
}
\value{
data.frame of results
}
\description{
Get server query results
}
