% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gor_query.R
\name{gorr__get_query_status}
\alias{gorr__get_query_status}
\title{Get query status from api. This is not a public function, but is called from \code{\link{gor_query}}}
\usage{
gorr__get_query_status(query_url, conn)
}
\arguments{
\item{query_url}{query url}

\item{conn}{connection object, see \code{\link{gor_connect}}}
}
\value{
response content object, see \code{\link[httr]{content}}
}
\description{
Get query status from api. This is not a public function, but is called from \code{\link{gor_query}}
}
