% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gor_query.R
\name{gorr__get_response_body}
\alias{gorr__get_response_body}
\title{Get response body, fail with an error if it has one}
\usage{
gorr__get_response_body(response, content.fun = httr::content)
}
\arguments{
\item{response}{from e.g.  \code{\link[httr]{POST}}, \code{\link[httr]{GET}}, \code{\link[httr]{DELETE}}}

\item{content.fun}{content function used to extract content from the response, e.g. \code{\link[httr]{text_content}}}
}
\value{
response body from (\code{\link[httr]{content}})
}
\description{
Get response body, fail with an error if it has one
}
