% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gor_query.R
\name{gorr__info}
\alias{gorr__info}
\title{Custom wrapper for message() with formated messages}
\usage{
gorr__info(msg, detail = NULL)
}
\arguments{
\item{msg}{message}

\item{detail}{exception details (chr or chr vector)}
}
\description{
Custom wrapper for message() with formated messages
}
