% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gor_query.R
\name{gorr__kill_query}
\alias{gorr__kill_query}
\title{Kill remote query}
\usage{
gorr__kill_query(query_url, conn)
}
\arguments{
\item{query_url}{query url}

\item{conn}{connection object, see \code{\link{platform_connect}}}
}
\description{
Kill remote query
}
