% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gor_query.R
\name{gorr__post_query}
\alias{gorr__post_query}
\title{Post query to api. This is not a public function, but is called from \code{\link{gor_query}}}
\usage{
gorr__post_query(query, conn, relations = NULL, persist = NULL)
}
\arguments{
\item{query}{GOR query}

\item{conn}{connection object, see \code{\link{platform_connect}}}

\item{relations}{data.frames to include with the query in the format \code{list(list(table_name = data.frame() ))}}

\item{persist}{remote path to file for saving results of the query into. Query results will not be fetched if this parameter is set.}
}
\value{
response content object, see \code{\link[httr]{content}}
}
\description{
Post query to api. This is not a public function, but is called from \code{\link{gor_query}}
}
