% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gor_queryserver.R
\name{gorr__queryserver}
\alias{gorr__queryserver}
\title{Get query from queryserver api. This is not a public function, but is called from \code{\link{gor_query}}}
\usage{
gorr__queryserver(query, conn, parse, relations, spinner = invisible, ...)
}
\arguments{
\item{query}{gor query string}

\item{conn}{connection object, see \code{\link{platform_connect}}}

\item{parse}{should output be parsed into a data.frame}

\item{relations}{list of tables to upload and make available in the query}

\item{spinner}{spinner function (e.g. some ascii animation for long running queries)}

\item{...}{additional arguments to be passed to \code{\link{gorr__post_query}}}
}
\value{
query results on text format
}
\description{
Get query from queryserver api. This is not a public function, but is called from \code{\link{gor_query}}
}
