% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gor_queryservice.R
\name{gorr__queryservice}
\alias{gorr__queryservice}
\title{Get query from queryservice api. This is not a public function, but is called from \code{\link{gor_query}}}
\usage{
gorr__queryservice(
  query,
  conn,
  page_size,
  parse,
  relations,
  persist,
  spinner = invisible,
  ...
)
}
\arguments{
\item{query}{gor query string}

\item{conn}{connection object, see \code{\link{platform_connect}}}

\item{page_size}{large results are returned in paged responses, this parameter controls the page size}

\item{parse}{should output be parsed into a data.frame}

\item{relations}{list of tables to upload and make available in the query}

\item{persist}{remote path to file for saving results of the query into. Query results will not be fetched if this parameter is set.}

\item{spinner}{spinner function (e.g. some ascii animation for long running queries)}

\item{...}{additional arguments to be passed to \code{\link{gorr__post_query}}}
}
\value{
query results on text format
}
\description{
Get query from queryservice api. This is not a public function, but is called from \code{\link{gor_query}}
}
