% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gor_connect.R
\name{gorr__reconnect}
\alias{gorr__reconnect}
\title{Reconnect using a connection object from \code{\link{gor_connect}}}
\usage{
gorr__reconnect(conn)
}
\arguments{
\item{conn}{connection object from \code{\link{gor_connect}}}
}
\value{
a new gor connection object (list)
}
\description{
Reconnect using a connection object from \code{\link{gor_connect}}
}
