% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gor_query.R
\name{gorr__warning}
\alias{gorr__warning}
\title{Custom wrapper for warning() with formatted error messages}
\usage{
gorr__warning(msg, detail = NULL, url = NULL)
}
\arguments{
\item{msg}{warning message}

\item{detail}{warning details (chr or chr vector)}

\item{url}{query url}
}
\description{
Custom wrapper for warning() with formatted error messages
}
