% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_matrix.R
\name{phemat_add_phenotype}
\alias{phemat_add_phenotype}
\title{Add a new phenotype to the matrix request.}
\usage{
phemat_add_phenotype(name, phenotype_matrix, missing_value = NA, label = NULL)
}
\arguments{
\item{name}{Phenotype name}

\item{phenotype_matrix}{A phenotype matrix structure}

\item{missing_value}{The string to substitute for a missing value in the data}

\item{label}{Optional label to apply to the phenotype}
}
\value{
a phenotype matrix object
}
\description{
Add a new phenotype to the matrix request.
}
\examples{
\dontrun{
phenotype_mat <- get_phenotype_matrix()
phenotype_mat <- phemat_add_phenotype(name="PT1", phenotype_matrix = phenotype_mat)
}
}
