% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_matrix_svc.R
\name{phemat_add_phenotypes}
\alias{phemat_add_phenotypes}
\title{Add a list of phenotypes to the matrix request.}
\usage{
phemat_add_phenotypes(names, phenotype_matrix, missing_value = NULL)
}
\arguments{
\item{names}{List of phenotype names}

\item{phenotype_matrix}{A phenotype matrix structure}

\item{missing_value}{The string to substitute for a missing value in the data}
}
\value{
a phenotype matrix object
}
\description{
Add a list of phenotypes to the matrix request.
}
\examples{
\dontrun{
phenotype_mat <- get_phenotype_matrix()
phenotype_mat <- phemat_add_phenotypes(name=list("PT1", "PT2"), phenotype_matrix = phenotype_mat)
}
}
