% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_matrix.R
\name{phemat_remove_phenotype}
\alias{phemat_remove_phenotype}
\title{Remove a phenotype from the matrix request.}
\usage{
phemat_remove_phenotype(name, phenotype_matrix)
}
\arguments{
\item{name}{phenotype names}

\item{phenotype_matrix}{A phenotype matrix structure}
}
\value{
a phenotype matrix object
}
\description{
Does not fail if the phenotype is not present.
}
\examples{
\dontrun{
phenotype_mat <- get_phenotype_matrix()
phenotype_mat <- phemat_add_phenotypes(name=list("PT1", "PT2"), phenotype_matrix = phenotype_mat)
phenotype_mat <- phemat_remove_phenotype(name= "PT1", phenotype_matrix = phenotype_mat)
}
}
