% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_update.R
\name{phenotype_add_tags}
\alias{phenotype_add_tags}
\title{Add new tag/s phenotype.}
\usage{
phenotype_add_tags(tags, phenotype)
}
\arguments{
\item{tags}{string or character vector of tag/s to be added. Tags should be seperated by comma e.g. "height,weight" or as vector "c("height", "weight")}

\item{phenotype}{phenotype structure, create or get it using \code{\link{create_phenotype}} or \code{\link{get_phenotype}}}
}
\value{
phenotype structure
}
\description{
Add new tag/s phenotype.
}
