% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-phenotype_connect.R
\name{phenotype_connect}
\alias{phenotype_connect}
\title{Set up a connection object, and call the phenotype-catalog endpoint to make sure everything is up and running on the
server side}
\usage{
phenotype_connect(api_key = NULL, project = NULL, root_url = NULL)
}
\arguments{
\item{api_key}{the api key from the /api-key-service endpoint of your CSA host through a browser}

\item{project}{project name}

\item{root_url}{root_url of the Query API to use, e.g. "http://localhost:1337". If left as NULL the function will try to get it from the environment variable GOR_API_ROOT_URL, otherwise it will take it from the iss part of the decoded JWT (api_key)}
}
\value{
returns a list with the connection data
}
\description{
This method is depricated and will be removed in future versions
}
\examples{
\dontrun{
api_key <- "...paste from your /api-key-service/token endpoint..."
conn <- phenotype_connect(api_key, "test_project")

# or have phenotype_connect take the api_key from the environment variable GOR_API_KEY:
conn <- phenotype_connect(project = "test_project")

}
}
