% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_services.R
\name{phenotype_get_error}
\alias{phenotype_get_error}
\title{Retrieve latest error from phenotype query run}
\usage{
phenotype_get_error(phenotype)
}
\arguments{
\item{phenotype}{phenotype structure, create or get it using \code{\link{create_phenotype}} or \code{\link{get_phenotype}}}
}
\value{
a formatted error message
}
\description{
Retrieve latest error from phenotype query run
}
\examples{
\dontrun{
api_key <- Sys.getenv("GOR_API_KEY")
project <- Sys.getenv("GOR_PROJECT")
conn <- platform_connect(api_key, project)
phenotype <- get_phenotype(name="height", conn)
tags <- phenotype_get_error(phenotype)
}
}
