% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_services.R
\name{phenotype_get_errors}
\alias{phenotype_get_errors}
\title{Retrieve all errors for phenotype query runs}
\usage{
phenotype_get_errors(phenotype)
}
\arguments{
\item{phenotype}{phenotype structure, create or get it using \code{\link{create_phenotype}} or \code{\link{get_phenotype}}}
}
\value{
a list of errors from query runs
}
\description{
Retrieve all errors for phenotype query runs
}
\examples{
\dontrun{
api_key <- Sys.getenv("GOR_API_KEY")
project <- Sys.getenv("GOR_PROJECT")
conn <- platform_connect(api_key, project)
phenotype <- get_phenotype(name="height", conn)
tags <- phenotype_get_errors(phenotype)
}
}
