% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_services.R
\name{phenotype_get_tags}
\alias{phenotype_get_tags}
\title{Retrieve all tags for phenotype}
\usage{
phenotype_get_tags(phenotype)
}
\arguments{
\item{phenotype}{phenotype structure, create or get it using \code{\link{create_phenotype}} or \code{\link{get_phenotype}}}
}
\value{
a list of phenotype tags
}
\description{
Retrieve all tags for phenotype
}
\examples{
\dontrun{
api_key <- Sys.getenv("GOR_API_KEY")
project <- Sys.getenv("GOR_PROJECT")
conn <- phenotype_connect(api_key, project)
name <- "height"
tags <- phenotype_get_tags(phenotype, conn)
}
}
