% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype.R
\name{phenotype_refresh}
\alias{phenotype_refresh}
\title{Refresh phenotype}
\usage{
phenotype_refresh(phenotype, conn = NULL)
}
\arguments{
\item{phenotype}{phenotype structure, create or get it using \code{\link{create_phenotype}} or \code{\link{get_phenotype}}}

\item{conn}{Deprecated : gor connection structure, create it using \code{\link{platform_connect}}}
}
\value{
phenotype structure
}
\description{
Fetches current (up-to-date) phenotype version from project.
}
