% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype.R
\name{phenotype_set_tags}
\alias{phenotype_set_tags}
\title{Set the tag/s for this phenotype, overriding previous tags}
\usage{
phenotype_set_tags(tags, phenotype, conn = NULL)
}
\arguments{
\item{tags}{string or character vector of tag/s to be added. Tags should be seperated by comma e.g. "height,weight" or as vector "c("height", "weight")}

\item{phenotype}{phenotype structure, create or get it using \code{\link{create_phenotype}} or \code{\link{get_phenotype}}}

\item{conn}{Deprecated : gor connection structure, create it using \code{\link{platform_connect}}}
}
\value{
phenotype structure
}
\description{
Set the tag/s for this phenotype, overriding previous tags
}
