% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_update.R
\name{phenotype_update}
\alias{phenotype_update}
\title{Update phenotype attributes}
\usage{
phenotype_update(
  phenotype,
  description = NULL,
  tags = list(),
  query = NULL,
  category = NULL,
  url = NULL
)
}
\arguments{
\item{phenotype}{phenotype structure, create or get it using \code{\link{get_phenotype}}}

\item{description}{phenotype description}

\item{tags}{comma separated string of tags eg. "height,weight", list or character vector "c("height", "weight")}

\item{query}{NOR query that defines this phenotype}

\item{category}{category for the phenotype (must be defined in the project - see get_categories)}

\item{url}{reference URL for the phenotype (to dataset or other reference)}
}
\value{
phenotype object
}
\description{
Update phenotype attributes
}
\examples{
\dontrun{
api_key <- Sys.getenv("GOR_API_KEY")
project <- Sys.getenv("GOR_PROJECT")
conn <- platform_connect(api_key, project)
name <- "height"
phenotype <- get_phenotype(name, conn)
description <- "individual height"
phenotype <- update_phenotype(phenotype, description=decription)
}
}
