% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_services.R
\name{phenotype_update_description}
\alias{phenotype_update_description}
\title{Update the phenotype with a new description}
\usage{
phenotype_update_description(description, phenotype, conn)
}
\arguments{
\item{description}{phenotype description}

\item{phenotype}{phenotype structure, create or get it using \code{\link{create_phenotype}} or \code{\link{get_phenotype}}}

\item{conn}{gor connection structure, create it using \code{\link{phenotype_connect}} or \code{\link{gor_connect}}}
}
\value{
an updated list with the phenotype object
}
\description{
Update the phenotype with a new description
}
\examples{
\dontrun{
api_key <- Sys.getenv("GOR_API_KEY")
project <- Sys.getenv("GOR_PROJECT")
conn <- phenotype_connect(api_key, project)
name <- "height"
phenotype <- get_phenotype(name, conn)
description <- "individual height"
phenotype <- update_phenotype_desc(description, phenotype, conn)
}
}
