% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_update.R
\name{phenotype_update_query}
\alias{phenotype_update_query}
\title{Update the phenotype with new query}
\usage{
phenotype_update_query(query, phenotype)
}
\arguments{
\item{query}{NOR query that defines this phenotype}

\item{phenotype}{phenotype structure, create or get it using \code{\link{get_phenotype}}}
}
\value{
phenotype structure
}
\description{
Update the phenotype with new query
}
