% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_playlist.R
\name{playlist_add_phenotype}
\alias{playlist_add_phenotype}
\title{Add phenotype/s to a playlist.}
\usage{
playlist_add_phenotype(name, playlist, conn)
}
\arguments{
\item{name}{comma seperated string of phenotypes existing in project to add eg. "pheno1,pheno2"  or character vector c("pheno1", "pheno2")}

\item{playlist}{playlist structure, create or get it using \code{\link{create_playlist}} or \code{\link{get_playlist}}}

\item{conn}{gor connection structure, create it using \code{\link{gor_connect}}}
}
\value{
A playlist object
}
\description{
Add phenotype/s to a playlist.
}
\examples{
\dontrun{
api_key <- Sys.getenv("GOR_API_KEY")
project <- Sys.getenv("GOR_PROJECT")
conn <- phenotype_connect(api_key, project)
name <- "testpl"
playlist <- create_playlist(name = name, conn)
playlist <- playlist_add_phenotype("pheno1", playlist, conn)
}
}
