% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_playlist.R
\name{playlist_add_phenotypes}
\alias{playlist_add_phenotypes}
\title{Add phenotype/s to playlist}
\usage{
playlist_add_phenotypes(names, playlist)
}
\arguments{
\item{names}{comma seperated string of phenotypes existing in project to add eg. "pheno1,pheno2" or character vector c("pheno1", "pheno2")}

\item{playlist}{playlist structure, create or get it using \code{\link{create_playlist}} or \code{\link{get_playlist}}}
}
\value{
A playlist object
}
\description{
Add phenotype/s to playlist
}
