% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_playlist.R
\name{playlist_delete}
\alias{playlist_delete}
\title{Delete a playlist from a project.}
\usage{
playlist_delete(playlist, conn = NULL)
}
\arguments{
\item{playlist}{playlist structure, create or get it using \code{\link{create_playlist}} or \code{\link{get_playlist}}}

\item{conn}{gor connection structure, create it using \code{\link{platform_connect}}}
}
\value{
A playlist object
}
\description{
Delete a playlist from a project.
}
\examples{
\dontrun{
api_key <- Sys.getenv("GOR_API_KEY")
project <- Sys.getenv("GOR_PROJECT")
conn <- platform_connect(api_key, project)
playlist <- get_playlist(conn, id=1)
playlist_delete(playlist, conn)
}
}
