% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_playlist.R
\name{playlist_delete_phenotypes}
\alias{playlist_delete_phenotypes}
\title{Delete phenotype/s from playlist}
\usage{
playlist_delete_phenotypes(names, playlist)
}
\arguments{
\item{names}{comma seperated string of phenotypes existing in project to delete eg. "pheno1,pheno2" or character vector c("pheno1", "pheno2")}

\item{playlist}{playlist structure, create or get it using \code{\link{create_playlist}} or \code{\link{get_playlist}}}
}
\value{
A playlist object
}
\description{
Delete phenotype/s from playlist
}
