% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_playlist.R
\name{playlist_update_description}
\alias{playlist_update_description}
\title{Update playlist's description}
\usage{
playlist_update_description(description, playlist)
}
\arguments{
\item{description}{playlist description}

\item{playlist}{playlist structure, create or get it using \code{\link{create_playlist}} or \code{\link{get_playlist}}}
}
\value{
A playlist object
}
\description{
Update playlist's description
}
