% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_playlist.R
\name{playlist_update_phenotypes}
\alias{playlist_update_phenotypes}
\title{Update playlist phenotypes}
\usage{
playlist_update_phenotypes(names, playlist, remove)
}
\arguments{
\item{names}{comma seperated string of phenotypes existing in project to add eg. "pheno1,pheno2" or character vector c("pheno1", "pheno2")}

\item{playlist}{playlist structure, create or get it using \code{\link{create_playlist}} or \code{\link{get_playlist}}}

\item{remove}{whether to delete the the phenotypes}
}
\value{
A playlist object
}
\description{
Update playlist phenotypes
}
