% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype.R
\name{plot__category}
\alias{plot__category}
\title{Plot categorical phenotypes. This is not a public function, but is called from \code{\link{phenotype_plot}}}
\usage{
plot__category(df, colname, fill)
}
\arguments{
\item{df}{phenotype data}

\item{colname}{name of column to plot}

\item{fill}{fill color}
}
\value{
ggplot2 object
}
\description{
Plot categorical phenotypes. This is not a public function, but is called from \code{\link{phenotype_plot}}
}
