% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_plot.R
\name{plot__set}
\alias{plot__set}
\title{Plot set phenotypes. This is not a public function, but is called from \code{\link{phenotype_plot}}}
\usage{
plot__set(df, colname, fill)
}
\arguments{
\item{df}{phenotype data}

\item{colname}{name of column to plot}

\item{fill}{fill color}
}
\value{
ggplot2 object
}
\description{
Plot set phenotypes. This is not a public function, but is called from \code{\link{phenotype_plot}}
}
