% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_services.R
\name{post_job}
\alias{post_job}
\title{Run a workflow job}
\usage{
post_job(pipeline_name, params, conn)
}
\arguments{
\item{pipeline_name}{Name of the pipeline to run}

\item{params}{Named list of parameters to forward to the job}

\item{conn}{connection object, see \code{\link{platform_connect}}}
}
\value{
WorkflowJob object
}
\description{
This is a low-level implementation on top of the workflow service's POST /jobs endpoint.
}
