% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_job.R
\name{workflow_check_status}
\alias{workflow_check_status}
\title{Get current status of job}
\usage{
workflow_check_status(job)
}
\arguments{
\item{job}{WorkflowJob object, see \code{\link{post_job}} and \code{\link{get_job}}}
}
\value{
Job status
}
\description{
Get current status of job
}
