% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_jobs.R
\name{workflow_refresh}
\alias{workflow_refresh}
\title{Refresh the local object of the serverside job object}
\usage{
workflow_refresh(job)
}
\arguments{
\item{job}{WorkflowJob object}
}
\value{
WorkflowJob object
}
\description{
Refresh the local object of the serverside job object
}
