% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_jobs.R
\name{workflow_wait}
\alias{workflow_wait}
\title{Wait for a running job to complete.}
\usage{
workflow_wait(job, max_seconds = NULL, poll_period = 0.5)
}
\arguments{
\item{job}{WorkflowJob object, see \code{\link{post_job}} and \code{\link{get_job}}}

\item{max_seconds}{raise an exception if the job runs longer than this}

\item{poll_period}{Number of seconds to wait between polling (max 10 seconds)}
}
\value{
WorkflowJob object
}
\description{
This is similar to the wait method in the Query Service and will wait by default
indefinitely for a workflow job to complete and poll the server regularly to update
the local status. When the job is completed (or failed) the method will return the
workflow job object.
}
