% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_job.R
\name{WorkflowJob}
\alias{WorkflowJob}
\title{WorkflowJob class constructor}
\usage{
WorkflowJob(response, conn)
}
\arguments{
\item{response}{workflow-api job response}

\item{conn}{connection object, see \code{\link{platform_connect}}}
}
\description{
Proxy object for a server side workflow job.
}
\details{
This object can be queried for the current status of a workflow job and will
automatically refresh from server until the job is finished (or failed)
}
