% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-catalog_services.R
\name{create_analysis_catalog}
\alias{create_analysis_catalog}
\title{Create a new analysis catalog in the current project}
\usage{
create_analysis_catalog(
  playlist_id,
  name,
  recipe_name,
  recipe_parameters,
  conn,
  covariate_phenotypes = NULL,
  excluded_pns = NULL
)
}
\arguments{
\item{playlist_id}{The id of the playlist that contains the phenotypes to use for analysis}

\item{name}{Analysis Catalog name}

\item{recipe_name}{The name of the recipe to use}

\item{recipe_parameters}{The parameters required to run the recipe}

\item{conn}{gor connection structure, create it using \code{\link{platform_connect}}}

\item{covariate_phenotypes}{The names of phenotypes to use as covariates, e.g. c("Pheno1','Pheno2')s (optional)}

\item{excluded_pns}{the PNs to exclude from the analysis, e.g. c('PN1','PN2') (optional)}
}
\value{
analysis_catalog structure
}
\description{
Create a new analysis catalog in the current project
}
