% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_matrix.R
\name{get_data}
\alias{get_data}
\title{Retrieve phenotype data or phenotype matrix data from the server.}
\usage{
get_data(phenotype_obj, conn)
}
\arguments{
\item{phenotype_obj}{phenotype or phenotype_matrix structure.}

\item{conn}{gor connection structure, create it using \code{\link{phenotype_connect}}}
}
\value{
tibble from server
}
\description{
If input is phenotype matrix object then phenotypes need te have been added to the phenotype matrix object. Create it
using \code{\link{get_phenotype_matrix}} and add to it
using \code{\link{phemat_add_phenotype}} or \code{\link{phemat_add_phenotypes}}
}
\examples{
\dontrun{
api_key <- Sys.getenv("GOR_API_KEY")
project <- Sys.getenv("GOR_PROJECT")
conn <- phenotype_connect(api_key, project)
phenotype_mat <- get_phenotype_matrix()
phenotype_mat <-  phemat_add_phenotypes(...)
phenotype_data <- get_data(phenotype_mat, conn)
}
}
