% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\alias{get_data.phenotype}
\alias{get_data.phenotype_matrix}
\alias{get_data.playlist}
\alias{get_data.phenotype_list}
\alias{get_data.default}
\title{Retrieve phenotype data or phenotype matrix data from the server.}
\usage{
get_data(pheno_obj, ...)

\method{get_data}{phenotype}(pheno_obj, conn = NULL, ...)

\method{get_data}{phenotype_matrix}(pheno_obj, conn = NULL, ...)

\method{get_data}{playlist}(pheno_obj, missing_value = NULL, base = NULL, ...)

\method{get_data}{phenotype_list}(pheno_obj, missing_value = NULL, base = NULL, ...)

\method{get_data}{default}(pheno_obj, conn, missing_value = NA, base = NULL, ...)
}
\arguments{
\item{pheno_obj}{phenotype names, phenotype, phenotype matrix or phenotype playlist structure.}

\item{...}{additional arguments}

\item{conn}{platform connection structure, create it using \code{\link{platform_connect}}}

\item{missing_value}{The string to substitute for a missing value in the data}

\item{base}{Optional name of base set}
}
\value{
tibble from server
}
\description{
If input is phenotype matrix object then phenotypes need te have been added to the phenotype matrix object. Create it
using \code{\link{get_phenotype_matrix}} and add to it
using \code{\link{phemat_add_phenotype}} or \code{\link{phemat_add_phenotypes}}
}
\section{Methods (by class)}{
\itemize{
\item \code{phenotype}: Get phenotype data

\item \code{phenotype_matrix}: Get data for phenotypes in phenotype matrix

\item \code{playlist}: Get data phenotypes in phenotype playlist

\item \code{phenotype_list}: Get data phenotypes in phenotype list

\item \code{default}: Get data phenotype/s by phenotype names
}}

\examples{
\dontrun{
api_key <- Sys.getenv("GOR_API_KEY")
project <- Sys.getenv("GOR_PROJECT")
conn <- platform_connect(api_key, project)
phenotype_mat <- get_phenotype_matrix()
phenotype_mat <-  phemat_add_phenotypes(...)
phenotype_data <- get_data(phenotype_mat, conn)
}
}
