% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_services.R
\name{get_phenotypes}
\alias{get_phenotypes}
\title{A list of all the phenotypes in the current project.}
\usage{
get_phenotypes(conn, tags = list(), limit = 100)
}
\arguments{
\item{conn}{gor connection structure, create it using \code{\link{gor_connect}}}

\item{tags}{Optional character, character vector or list of tags to filter for.}

\item{limit}{Maximum number of results (default: 100)}
}
\value{
List of phenotypes
}
\description{
A list of all the phenotypes in the current project.
}
\examples{
\dontrun{
api_key <- Sys.getenv("GOR_API_KEY")
project <- Sys.getenv("GOR_PROJECT")
conn <- phenotype_connect(api_key, project)
phenotypes <- get_phenotypes(conn)
}
}
