% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gor_create.R
\name{gor_create}
\alias{gor_create}
\title{gor_create}
\usage{
gor_create(
  ...,
  defs = NULL,
  conn = NULL,
  replace = NULL,
  query.service = "queryserver"
)
}
\arguments{
\item{...}{ellipse for the relations to include}

\item{defs}{definitions string}

\item{conn}{connection object from \code{\link{platform_connect}}}

\item{replace}{replace a previously created \code{gor_create} closure. When supplied, the two will be merged, overwriting existing values with the current values if they have been previously defined}

\item{query.service}{query service to use - either 'queryservice' (old) or 'queryserver' (new). Default: queryservice}
}
\value{
partial-application of function \code{\link{gor_query}} with the \code{conn} and \code{relations} parameters set.
}
\description{
Creates a partial-application of \code{\link{gor_query}} in the context of \code{conn} and the listed relations in the \code{...} arguments
}
\details{
Relations can be either local dataframes (uploaded) or string queries(turned into create statements and prepended to the query when executed).
}
\examples{
\dontrun{
query <- gor_create(conn = conn)
query <- gor_create(air = airquality, replace = query)
query("nor [air] | where Month=8")
months_df <- data.frame(id = 1:12, MonthName = month.abb)
query <- gor_create(months = months_df, replace = query)
query("nor [air] | map -c Month [months]")
}
}
