% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{gorr}
\alias{gorr}
\alias{gorr-package}
\title{GOR-R: R library for remote execution of GOR queries and authorization with WuXi NextCode gateway API.}
\description{
GOR-R: R library for remote execution of GOR queries and authorization with WuXi NextCode gateway API.
}
\section{GOR-R functions}{

The main functions you'll be using are:
\itemize{
\item \code{\link{gor_connect}} to establish a connection to your GOR API, and
\item \code{\link{gor_query}} to run your queries
}
}

\examples{
\dontrun{
library(gorr)
# Go to your CSA instance of choice that runs the Query API, e.g.
# https://your-wxnc-instance
# Note the internal names of the projects you have access to
# go to and copy your api_key from: https://your-wxnc-instance/api-key-service/token
# Paste it into the variable below:

api_key <- ""

# Make connection object
conn <- gor_connect(api_key, "your_project")

# Print the connection details
print(conn)

# Run a simple query

results <- gor_query("gor #dbsnp# | top 10000", conn)

print(results)
}
}
