% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_matrix.R
\name{phemat_add_phenotypes}
\alias{phemat_add_phenotypes}
\title{Add multiple phenotypes to a phenotype matrix request.}
\usage{
phemat_add_phenotypes(names, phenotype_matrix, missing_value = NA)
}
\arguments{
\item{names}{a comma seperated string or character vector of phenotype/s to be added. E.g. "height,weight" or c("height", "weight")}

\item{phenotype_matrix}{A phenotype matrix structure}

\item{missing_value}{The string to substitute for missing values in the data}
}
\value{
a phenotype matrix object
}
\description{
Add multiple phenotypes to a phenotype matrix request.
}
\examples{
\dontrun{
phenotype_mat <- get_phenotype_matrix()
phenotype_mat <- phemat_add_phenotypes(names=c("PT1", "PT2"), phenotype_matrix = phenotype_mat)
}
}
