% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_svc.R
\name{phenotype_add_tag}
\alias{phenotype_add_tag}
\title{Add a new tag/s to this phenotype.}
\usage{
phenotype_add_tag(tag, phenotype, conn)
}
\arguments{
\item{tag}{string of tag/s to be added. Multiple tags should be seperated by comma e.g. "height,weight" or character vector "c("height", "weight")}

\item{phenotype}{phenotype structure, create or get it using \code{\link{create_phenotype}} or \code{\link{get_phenotype}}}

\item{conn}{gor connection structure, create it using \code{\link{phenotype_connect}} or \code{\link{gor_connect}}}
}
\value{
an updated phenotype object
}
\description{
Add a new tag/s to this phenotype.
}
