% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype.R
\name{phenotype_delete}
\alias{phenotype_delete}
\title{Delete a phenotype, including all data from a project}
\usage{
phenotype_delete(phenotype, conn = NULL)
}
\arguments{
\item{phenotype}{phenotype structure, create or get it using \code{\link{create_phenotype}} or \code{\link{get_phenotype}}}

\item{conn}{Deprecated : gor connection structure, create it using \code{\link{platform_connect}}}
}
\description{
Delete a phenotype, including all data from a project
}
\examples{
\dontrun{
api_key <- Sys.getenv("GOR_API_KEY")
project <- Sys.getenv("GOR_PROJECT")
conn <- platform_connect(api_key, project)
name <- "height"
phenotype <- get_phenotype(name, conn)
phenotype_delete(phenotype, conn)
}
}
