% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_deprecated.R
\name{phenotype_delete_tag}
\alias{phenotype_delete_tag}
\title{Delete a tag/s from phenotype.}
\usage{
phenotype_delete_tag(tag, phenotype, conn = NULL)
}
\arguments{
\item{tag}{string of tag/s to be added seperated by comma eg. "height,weight"  or character vector "c("height", "weight")}

\item{phenotype}{phenotype structure, create or get it using \code{\link{create_phenotype}} or \code{\link{get_phenotype}}}

\item{conn}{Deprecated : gor connection structure, create it using \code{\link{platform_connect}}}
}
\value{
an updated list with the phenotype object
}
\description{
This method is depricated and will be removed in future versions. use \code{\link{phenotype_delete_tags}} instead.
}
