% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_services.R
\name{phenotype_refresh}
\alias{phenotype_refresh}
\title{Refresh phenotype}
\usage{
phenotype_refresh(phenotype, conn)
}
\arguments{
\item{phenotype}{phenotype structure, create or get it using \code{\link{create_phenotype}} or \code{\link{get_phenotype}}}

\item{conn}{gor connection structure, create it using \code{\link{platform_connect}}}
}
\value{
a list with the phenotype object
}
\description{
Refresh phenotype
}
\examples{
\dontrun{
api_key <- Sys.getenv("GOR_API_KEY")
project <- Sys.getenv("GOR_PROJECT")
conn <- platform_connect(api_key, project)
name <- "height"
result_type <- "QT"
description <- "Height of individuals"
phenotype <- create_phenotype(name, result_type, conn, description)
phenotype <- phenotype_refresh(phenotype, conn)
}
}
