% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_update.R
\name{phenotype_update_description}
\alias{phenotype_update_description}
\title{Update the phenotype with a new description}
\usage{
phenotype_update_description(description, phenotype, conn = NULL)
}
\arguments{
\item{description}{phenotype description}

\item{phenotype}{phenotype structure, create or get it using \code{\link{get_phenotype}}}

\item{conn}{Deprecated : gor connection structure, create it using \code{\link{phenotype_connect}} or \code{\link{platform_connect}}}
}
\value{
phenotype structure
}
\description{
Update the phenotype with a new description
}
