% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype.R
\name{phenotype_upload_data}
\alias{phenotype_upload_data}
\title{Upload phenotype data}
\usage{
phenotype_upload_data(phenotype, data, conn = NULL)
}
\arguments{
\item{phenotype}{phenotype structure, create or get it using \code{\link{create_phenotype}} or \code{\link{get_phenotype}}}

\item{data}{a data.frame or list of lists to be uploaded
\itemize{
\item For phenotypes of result_type 'SET' - data should be either a list single column data.frame.
\item For other phenotypes the data.frame/tibble should contain 2 columns, pn and value or if list a pn value pair for each pn.
}}

\item{conn}{Deprecated : gor connection structure, create it using \code{\link{platform_connect}}}
}
\description{
The data is expected to be a data.frame, tibble or list of lists. If data.frame or tibble is provided it should contain 2 columns.
and list of lists should be of the form \code{list(list(pn1, value1),list(pn2,value2))} or \code{list(list(pn1),list(pn2))} For phenotypes of result_type SET.
The \code{result_type} of the phenotype dictates if each sublist should contain one or two items.
}
