% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_services.R
\name{phenotype_upload_data}
\alias{phenotype_upload_data}
\title{Upload phenotype data}
\usage{
phenotype_upload_data(phenotype, data, conn)
}
\arguments{
\item{phenotype}{phenotype structure, create or get it using \code{\link{create_phenotype}} or \code{\link{get_phenotype}}}

\item{data}{a list of lists to be uploaded}

\item{conn}{gor connection structure, create it using \code{\link{platform_connect}}}
}
\description{
The data is expected to be a list of lists.
e.g. \code{data <- list(list('a'), list('b'))}.
The \code{result_type} of the phenotype dictates
if each sublist should contain one or two items.
}
