% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_playlist.R
\name{playlist_add_phenotype}
\alias{playlist_add_phenotype}
\title{Add phenotype to a playlist.}
\usage{
playlist_add_phenotype(name, playlist, conn = NULL)
}
\arguments{
\item{name}{name of phenotype existing in project}

\item{playlist}{playlist structure, create or get it using \code{\link{create_playlist}} or \code{\link{get_playlist}}}

\item{conn}{gor connection structure, create it using \code{\link{platform_connect}}}
}
\value{
A playlist object
}
\description{
For adding multiple phenotypes to playlist see: or \code{\link{playlist_add_phenotypes}}
}
\examples{
\dontrun{
api_key <- Sys.getenv("GOR_API_KEY")
project <- Sys.getenv("GOR_PROJECT")
conn <- platform_connect(api_key, project)
name <- "testpl"
playlist <- create_playlist(name = name, conn)
playlist <- playlist_add_phenotype("pheno1", playlist, conn)
}
}
